Consider override
Fixes#73
