+++ src/main/java/com/treasure/hunt/view/MainController.java
@@ -2,8 +2,6 @@ package com.treasure.hunt.view;
 
 import com.treasure.hunt.game.GameEngine;
 import com.treasure.hunt.game.GameManager;
-import com.treasure.hunt.jts.AdvancedShapeWriter;
-import com.treasure.hunt.jts.PointTransformation;
 import com.treasure.hunt.strategy.hider.Hider;
 import com.treasure.hunt.strategy.searcher.Searcher;
 import com.treasure.hunt.utils.ReflectionUtils;
@@ -18,20 +16,15 @@ import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
 import javafx.fxml.FXML;
 import javafx.scene.Node;
-import javafx.scene.canvas.Canvas;
 import javafx.scene.control.Button;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.Label;
 import javafx.scene.control.SplitPane;
-import javafx.scene.input.MouseEvent;
-import javafx.scene.input.ScrollEvent;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.VBox;
 import javafx.util.StringConverter;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
-import org.jfree.fx.FXGraphics2D;
-import org.locationtech.jts.math.Vector2D;
 import org.reflections.Reflections;
 
 import java.lang.reflect.Modifier;
@@ -45,8 +38,6 @@ import java.util.stream.Collectors;
 @Slf4j
 public class MainController {
 
-    public Canvas canvas;
-    public Pane canvasPane;
     public SplitPane mainSplitPane;
 
     public Pane leftWidgetBar;
@@ -54,7 +45,9 @@ public class MainController {
 
     public VBox rightToolbar;
     public VBox leftToolbar;
-
+    @FXML
+    public CanvasController canvasController;
+    public Pane canvas;
     @FXML
     private WidgetBarController leftWidgetBarController;
     @FXML
@@ -77,20 +70,14 @@ public class MainController {
     @Getter
     private final ObjectProperty<GameManager> gameManager = new SimpleObjectProperty<>();
 
-    private PointTransformation transformation = new PointTransformation();
-    private AdvancedShapeWriter shapeWriter = new AdvancedShapeWriter(transformation);
 
-    private FXGraphics2D graphics2D;
-
-    private Vector2D dragStart = new Vector2D();
-    private Vector2D offsetBackup = new Vector2D();
 
     public void initialize() {
+        canvasController.setGameManager(gameManager);
         setListStringConverters();
         fillLists();
         addPromptBindings();
         bindStartButtonState();
-        makeCanvasResizable();
         addToolbarStyleClasses();
         bindWidgetBarVisibility();
     }
@@ -297,21 +284,20 @@ public class MainController {
         }
 
         if (initialize) {
-            graphics2D = new FXGraphics2D(canvas.getGraphicsContext2D());
-            gameManager.addListener(change -> drawShapes());
+            gameManager.addListener(change -> canvasController.drawShapes());
             initGameUI();
         }
     }
 
     public void initGameUI() {
-        drawShapes();
+        canvasController.drawShapes();
         addTreasureInspector();
         nextButton.setDisable(false);
     }
 
     public void previousButtonClicked() {
         gameManager.get().previous();
-        drawShapes();
+        canvasController.drawShapes();
         if (gameManager.get().isFirstStepShown()) {
             previousButton.setDisable(true);
         }
@@ -320,67 +306,11 @@ public class MainController {
 
     public void nextButtonClicked() {
         gameManager.get().next();
-        drawShapes();
+        canvasController.drawShapes();
         if (gameManager.get().isGameFinished() && gameManager.get().isSimStepLatest()) {
             nextButton.setDisable(true);
             logLabel.setText("Game ended");
         }
         previousButton.setDisable(false);
-    }
-
-    public void makeCanvasResizable() {
-        canvas.widthProperty().addListener((observableValue, number, t1) -> drawShapes());
-
-        canvas.widthProperty().addListener((observableValue, number, t1) -> drawShapes());
-
-        canvas.heightProperty().bind(canvasPane.heightProperty());
-        canvas.widthProperty().bind(canvasPane.widthProperty());
-    }
-
-    private void drawShapes() {
-        if (gameManager.isNotNull().get()) {
-            deleteShapes();
-            gameManager.get().getGeometryItems().forEach(geometryItem ->
-                    geometryItem.draw(graphics2D, shapeWriter)
-            );
-        }
-    }
-
-    private void deleteShapes() {
-        canvas.getGraphicsContext2D().clearRect(0, 0, canvas.getWidth(), canvas.getHeight());
-    }
-
-    public void onCanvasClicked(MouseEvent mouseEvent) {
-        if (gameManager == null) {
-            return;
-        }
-        offsetBackup = transformation.getOffset();
-        dragStart = Vector2D.create(mouseEvent.getX(), mouseEvent.getY());
-    }
-
-    public void onCanvasDragged(MouseEvent mouseEvent) {
-        if (gameManager == null) {
-            return;
-        }
-        Vector2D dragOffset = Vector2D.create(mouseEvent.getX(), mouseEvent.getY()).subtract(dragStart);
-        transformation.setOffset(dragOffset.add(offsetBackup));
-        drawShapes();
-    }
-
-    public void onCanvasZoom(ScrollEvent scrollEvent) {
-        if (gameManager == null) {
-            return;
-        }
-        Vector2D mouse = new Vector2D(scrollEvent.getX(), scrollEvent.getY());
-        Vector2D direction = transformation.getOffset().subtract(mouse);
-
-        double oldScale = transformation.getScale();
-        double newScale = oldScale * Math.exp(scrollEvent.getDeltaY() * 1e-2);
-
-        if (newScale > 0) {
-            transformation.setScale(newScale);
-            transformation.setOffset(mouse.add(direction.multiply(newScale / oldScale)));
-            drawShapes();
-        }
     }
 }
