+++ src/main/resources/layout/canvas.fxml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.canvas.Canvas?>
+<?import javafx.scene.layout.Pane?>
+<Pane xmlns:fx="http://javafx.com/fxml" id="canvasPane" fx:id="canvasPane" prefHeight="200.0" prefWidth="200.0"
+      xmlns="http://javafx.com/javafx"
+      fx:controller="com.treasure.hunt.view.CanvasController">
+    <children>
+        <Canvas
+                id="gameCanvas"
+                fx:id="canvas"
+                onMouseDragged="#onCanvasDragged"
+                onMousePressed="#onCanvasClicked"
+                onScroll="#onCanvasZoom"
+        />
+    </children>
+</Pane>
