+++ src/test/com/treasure/hunt/strategy/hider/implementations/MaxAreaAngularHintStrategyTest.java
